Do not rely on the index.html template being in the static directory.
First, index.html will be in the *template* directory, not the static directory. Second, we don’t know for sure what files will be in the static directory. A check like this should be done at the level of the specific app building the static resources (like the jupyter lab app itself).